update eval results with fake quant (weight_quant=per_channel, act_quant=per_token) LLM.int8()

+++ examples/export_int8_model.py
@@ -31,31 +31,27 @@ if __name__ == '__main__':
     parser.add_argument('--export-FT', default=False, action="store_true")
     args = parser.parse_args()
     
-    # model = OPTForCausalLM.from_pretrained(
-    #     args.model_name, device_map="auto", torch_dtype=torch.float16)
-
-    model = AutoModelForCausalLM.from_pretrained(
+    model = OPTForCausalLM.from_pretrained(
         args.model_name, device_map="auto", torch_dtype=torch.float16)
-
     act_scales = torch.load(args.act_scales)
-    # smooth_lm(model, act_scales, 0.5)
-    smooth_lm(model, act_scales, 0.85)
-    #tokenizer = AutoTokenizer.from_pretrained(args.model_name)
-    tokenizer = LlamaTokenizer.from_pretrained(args.model_name, model_max_length=512)
+    smooth_lm(model, act_scales, 0.5)
+    tokenizer = LlamaTokenizer.from_pretrained(args.model_name)
     tokenizer.pad_token_id = tokenizer.eos_token_id
 
-    # if not os.path.exists(args.dataset_path):
-    #     print(f'Cannot find the dataset at {args.dataset_path}')
-    #     print('Please download the Pile dataset and put the validation set at the path')
-    #     print('You can download the validation dataset of the Pile at https://mystic.the-eye.eu/public/AI/pile/val.jsonl.zst')
-    #     raise FileNotFoundError
+
+    if not os.path.exists(args.dataset_path):
+        print(f'Cannot find the dataset at {args.dataset_path}')
+        print('Please download the Pile dataset and put the validation set at the path')
+        print('You can download the validation dataset of the Pile at https://mystic.the-eye.eu/public/AI/pile/val.jsonl.zst')
+        raise FileNotFoundError
 
     decoder_layer_scales, raw_scales = get_static_decoder_layer_scales(model,
                                                                        tokenizer,
                                                                        args.dataset_path,
                                                                        num_samples=args.num_samples,
                                                                        seq_len=args.seq_len)
-    output_path = Path(args.output_path) / (Path(args.model_name).name + "-smoothquant.pt")
+    output_path = Path(args.output_path) / (Path(args.model_name).name + "-smoothquant")
+
     if args.export_FT:
         model.save_pretrained(output_path)
         print(f"Saved smoothed model at {output_path}")
